+++ pycosio/_core/functions_os_path.py
@@ -1,9 +1,8 @@
 # coding=utf-8
 """Cloud object compatibles standard library 'os.path' equivalent functions"""
 import os
-from os.path import relpath as os_path_relpath
+from os.path import relpath as os_path_relpath, samefile as os_path_samefile
 
-from pycosio._core.compat import samefile as os_path_samefile
 from pycosio._core.storage_manager import get_instance
 from pycosio._core.functions_core import equivalent_to, format_and_is_storage
 from pycosio._core.exceptions import handle_os_exceptions