+++ pycosio/_core/io_base.py
@@ -3,10 +3,9 @@
 from concurrent.futures import ThreadPoolExecutor
 from functools import wraps
 from io import IOBase, UnsupportedOperation
-from threading import Lock
 from itertools import chain
-
-from pycosio._core.compat import fsdecode
+from os import fsdecode
+from threading import Lock
 
 
 class ObjectIOBase(IOBase):