+++ pycosio/storage/azure_blob/_base_blob.py
@@ -1,7 +1,5 @@
 # coding=utf-8
 """Microsoft Azure Blobs Storage: Base for all blob types"""
-from __future__ import absolute_import  # Python 2: Fix azure import
-
 from io import IOBase
 
 from pycosio._core.io_base import memoizedmethod