+++ pycosio/storage/azure_blob/_block_blob.py
@@ -1,7 +1,5 @@
 # coding=utf-8
 """Microsoft Azure Blobs Storage: Block blobs"""
-from __future__ import absolute_import  # Python 2: Fix azure import
-
 from random import choice
 from string import ascii_lowercase as _ascii_lowercase