+++ pycosio/storage/azure_blob/_page_blob.py
@@ -1,7 +1,5 @@
 # coding=utf-8
 """Microsoft Azure Blobs Storage: Pages blobs"""
-from __future__ import absolute_import  # Python 2: Fix azure import
-
 from os import SEEK_SET, SEEK_END
 
 from azure.storage.blob import PageBlobService