+++ airfs/shutil.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+"""Standard library "shutil" equivalents"""
+
+from shutil import *  # noqa
+from airfs._core.functions_shutil import copy, copyfile  # noqa
+
+import shutil as _src_module
+__all__ = _src_module.__all__
+del _src_module