+++ tests/test_storage_azure_file.py
@@ -1,5 +1,9 @@
 # coding=utf-8
 """Test airfs.storage.azure_file"""
+import pytest
+
+pytest.importorskip("azure.storage.file")
+
 UNSUPPORTED_OPERATIONS = (
     'symlink',