+++ tests/test_storage_swift.py
@@ -2,6 +2,8 @@
 """Test airfs.storage.swift"""
 import pytest
 
+pytest.importorskip("swiftclient")
+
 UNSUPPORTED_OPERATIONS = (
     'symlink',